requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.4
django-cors-headers==4.2.0
django-filter==23.2
django-phonenumber-field==7.1.0
django-phonenumbers==1.0.1
django-postgres-extensions==0.9.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
phonenumbers==8.13.18
psycopg2-binary==2.9.7
PyJWT==2.8.0
python-dotenv==1.0.0
python-http-client==3.3.7
pytz==2023.3
requests==2.31.0
sendgrid==6.10.0
six==1.16.0
sqlparse==0.4.4
starkbank-ecdsa==2.2.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
